refactor: Replace server-side magit.status with generic commands.run_parallel
Move command selection from the Rust server to the Elisp client,
eliminating the fragile build-process-cache translation layer that
had to reconstruct exact magit argument patterns.

Server: Replace magit.rs (614 lines) with commands.rs (182 lines)
containing a generic run_parallel handler that takes [{key, cmd,
args, cwd}] and runs them in parallel via thread::scope.

Client: Remove tramp-rpc-magit-status, build-process-cache (~280
lines of key reconstruction), and state-file-exists-p. Replace with
prefetch-git-commands that defines ~39 git commands + 34 state file
checks, all sent in one RPC call. Results stored directly as the
process-file cache with no reconstruction needed.

State file existence checks (.git/MERGE_HEAD etc.) are included in
the parallel batch as test -e commands and cached in file-exists-cache.
